+++ src/main/kotlin/net/pwall/log/LoggerUtil.kt
@@ -2,7 +2,7 @@
  * @(#) LoggerUtil.kt
  *
  * log-front-kotlin  Logging interface in Kotlin
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -28,6 +28,100 @@ package net.pwall.log
 import kotlin.reflect.KClass
 
 /**
- * Get [Logger] for a Kotlin [KClass] (Kotlin equivalent of same function taking Java [Class]).
+ * Get a [Logger] for a Kotlin [KClass] (Kotlin equivalent of same function taking a Java [Class]).
+ *
+ * @receiver            [LoggerFactory]
+ * @param   kotlinClass a Kotlin [KClass]
+ * @return              a [Logger]
  */
 fun LoggerFactory.getLogger(kotlinClass: KClass<*>): Logger = getLogger(kotlinClass.qualifiedName)
+
+/**
+ * Test whether a [LogItem] represents a message with level `TRACE` and the specified text.
+ *
+ * @receiver        a [LogItem]
+ * @param   text    the text to be compared
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isTrace(text: String): Boolean = level == Level.TRACE && this.text == text
+
+/**
+ * Test whether a [LogItem] represents a message with level `TRACE` and the specified text.
+ *
+ * @receiver        a [LogItem]
+ * @param   regex   the [Regex]
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isTrace(regex: Regex): Boolean = level == Level.TRACE && regex.containsMatchIn(text)
+
+/**
+ * Test whether a [LogItem] represents a message with level `DEBUG` and the specified text.
+ *
+ * @receiver        a [LogItem]
+ * @param   text    the text to be compared
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isDebug(text: String): Boolean = level == Level.DEBUG && this.text == text
+
+/**
+ * Test whether a [LogItem] represents a message with level `DEBUG` and the specified text.
+ *
+ * @receiver        a [LogItem]
+ * @param   regex   the [Regex]
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isDebug(regex: Regex): Boolean = level == Level.DEBUG && regex.containsMatchIn(text)
+
+/**
+ * Test whether a [LogItem] represents a message with level `INFO` and the specified text.
+ *
+ * @receiver        a [LogItem]
+ * @param   text    the text to be compared
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isInfo(text: String): Boolean = level == Level.INFO && this.text == text
+
+/**
+ * Test whether a [LogItem] represents a message with level `INFO` and text that matches the given [Regex].
+ *
+ * @receiver        a [LogItem]
+ * @param   regex   the [Regex]
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isInfo(regex: Regex): Boolean = level == Level.INFO && regex.containsMatchIn(text)
+
+/**
+ * Test whether a [LogItem] represents a message with level `WARN` and the specified text.
+ *
+ * @receiver        a [LogItem]
+ * @param   text    the text to be compared
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isWarning(text: String): Boolean = level == Level.WARN && this.text == text
+
+/**
+ * Test whether a [LogItem] represents a message with level `WARN` and text that matches the given [Regex].
+ *
+ * @receiver        a [LogItem]
+ * @param   regex   the [Regex]
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isWarning(regex: Regex): Boolean = level == Level.WARN && regex.containsMatchIn(text)
+
+/**
+ * Test whether a [LogItem] represents a message with level `ERROR` and the specified text.
+ *
+ * @receiver        a [LogItem]
+ * @param   text    the text to be compared
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isError(text: String): Boolean = level == Level.ERROR && this.text == text
+
+/**
+ * Test whether a [LogItem] represents a message with level `ERROR` and text that matches the given [Regex].
+ *
+ * @receiver        a [LogItem]
+ * @param   regex   the [Regex]
+ * @return          `true` if the [LogItem] matches
+ */
+infix fun LogItem.isError(regex: Regex): Boolean = level == Level.ERROR && regex.containsMatchIn(text)