+++ src/test/kotlin/net/pwall/log/LoggerUtilTest.kt
@@ -26,12 +26,60 @@
 package net.pwall.log
 
 import kotlin.test.Test
+import kotlin.test.assertFalse
+import kotlin.test.assertTrue
 import kotlin.test.expect
+import java.time.Instant
 
 class LoggerUtilTest {
 
     @Test fun `should get logger for Kotlin class`() {
         expect("net.pwall.log.LoggerUtilTest") { LoggerFactory.getDefault().getLogger(LoggerUtilTest::class).name }
+    }
+
+    @Test fun `should check whether LogItem is trace`() {
+        val logItem = LogItem(Instant.now(), "xyz", Level.TRACE, "Hello", null)
+        assertTrue { logItem isTrace "Hello" }
+        assertTrue { logItem isTrace Regex("^H") }
+        assertFalse { logItem isDebug "Hello" }
+        assertFalse { logItem isTrace "Goodbye" }
+        assertFalse { logItem isTrace Regex("^Hello world") }
+    }
+
+    @Test fun `should check whether LogItem is debug`() {
+        val logItem = LogItem(Instant.now(), "xyz", Level.DEBUG, "Hello", null)
+        assertTrue { logItem isDebug "Hello" }
+        assertTrue { logItem isDebug Regex("^H") }
+        assertFalse { logItem isInfo "Hello" }
+        assertFalse { logItem isDebug "Goodbye" }
+        assertFalse { logItem isDebug Regex("^Hello world") }
+    }
+
+    @Test fun `should check whether LogItem is info`() {
+        val logItem = LogItem(Instant.now(), "xyz", Level.INFO, "Hello", null)
+        assertTrue { logItem isInfo "Hello" }
+        assertTrue { logItem isInfo Regex("^H") }
+        assertFalse { logItem isWarning "Hello" }
+        assertFalse { logItem isInfo "Goodbye" }
+        assertFalse { logItem isInfo Regex("^Hello world") }
+    }
+
+    @Test fun `should check whether LogItem is warning`() {
+        val logItem = LogItem(Instant.now(), "xyz", Level.WARN, "Hello", null)
+        assertTrue { logItem isWarning "Hello" }
+        assertTrue { logItem isWarning Regex("^H") }
+        assertFalse { logItem isError "Hello" }
+        assertFalse { logItem isWarning "Goodbye" }
+        assertFalse { logItem isWarning Regex("^Hello world") }
+    }
+
+    @Test fun `should check whether LogItem is error`() {
+        val logItem = LogItem(Instant.now(), "xyz", Level.ERROR, "Hello", null)
+        assertTrue { logItem isError "Hello" }
+        assertTrue { logItem isError Regex("^H") }
+        assertFalse { logItem isWarning  "Hello" }
+        assertFalse { logItem isError "Goodbye" }
+        assertFalse { logItem isError Regex("^Hello world") }
     }
 
 }